requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
gunicorn>=21.2.0 